requirements.txt
python-dotenv~=0.19.2
prometheus_client~=0.12.0
mcstatus~=7.0.0